[Added new Interact tooltip for better performance]
