+++ Assets/Prefabs/Environment/Trees/Tree Orange 8.prefab
@@ -112,6 +112,8 @@ MonoBehaviour:
   isFocus: 0
   hasInteracted: 0
   player: {fileID: 0}
+  checkScale: 0
+  maxScale: 0
   drops:
   - item: {fileID: 11400000, guid: dc0cd158442ec9b4a9e2ff201649af1c, type: 2}
     amount: 4
@@ -134,19 +136,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: c4cd36f15bf750343b472cd9e8c41d61, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  highlightColor: {r: 1, g: 1, b: 1, a: 1}
-  highlightSpeed: 4
-  hideTooltip: 0
-  followingMouseCursor: 1
-  tooltipPosition: {x: 0, y: 0}
-  tooltipText: 
-  tooltipColor: {r: 1, g: 1, b: 1, a: 1}
-  tooltipSize: 1
-  tooltipFont: {fileID: 0}
-  tooltipAlignment: 0
-  interactIcon: {fileID: 2800000, guid: 1e26e85a575ab744bae883b76425b1d0, type: 3}
-  iconSize: 40
-  iconOffset: 25
+  mouseCursor: {fileID: 2800000, guid: dfd1ebb9afa91404b8fc0d2af6df5d1d, type: 3}
 --- !u!1 &9008696042882691246
 GameObject:
   m_ObjectHideFlags: 0
