+++ Assets/Prefabs/Managers/CursorManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 32d964d77ab251141bafdf6b146a0285
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
