+++ Assets/Prefabs/ScenePrefabs/=====_ Managers.prefab
@@ -50,6 +50,7 @@ Transform:
   - {fileID: 7209149007321884530}
   - {fileID: 440583191854263974}
   - {fileID: 1272368521873471885}
+  - {fileID: 1792154953810459228}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -799,4 +800,79 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 6003750396554747860, guid: ac9a21b2fbaebf14188115954389de9f,
     type: 3}
   m_PrefabInstance: {fileID: 6146533027607130482}
+  m_PrefabAsset: {fileID: 0}
+--- !u!1001 &7019181396866707800
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 1854814581397351230}
+    m_Modifications:
+    - target: {fileID: 3546767374556701778, guid: 32d964d77ab251141bafdf6b146a0285,
+        type: 3}
+      propertyPath: m_Name
+      value: CursorManager
+      objectReference: {fileID: 0}
+    - target: {fileID: 8770240368010660612, guid: 32d964d77ab251141bafdf6b146a0285,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8770240368010660612, guid: 32d964d77ab251141bafdf6b146a0285,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8770240368010660612, guid: 32d964d77ab251141bafdf6b146a0285,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8770240368010660612, guid: 32d964d77ab251141bafdf6b146a0285,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8770240368010660612, guid: 32d964d77ab251141bafdf6b146a0285,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8770240368010660612, guid: 32d964d77ab251141bafdf6b146a0285,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8770240368010660612, guid: 32d964d77ab251141bafdf6b146a0285,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 8770240368010660612, guid: 32d964d77ab251141bafdf6b146a0285,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 9
+      objectReference: {fileID: 0}
+    - target: {fileID: 8770240368010660612, guid: 32d964d77ab251141bafdf6b146a0285,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8770240368010660612, guid: 32d964d77ab251141bafdf6b146a0285,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8770240368010660612, guid: 32d964d77ab251141bafdf6b146a0285,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 32d964d77ab251141bafdf6b146a0285, type: 3}
+--- !u!4 &1792154953810459228 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 8770240368010660612, guid: 32d964d77ab251141bafdf6b146a0285,
+    type: 3}
+  m_PrefabInstance: {fileID: 7019181396866707800}
   m_PrefabAsset: {fileID: 0}
