+++ Assets/Scripts/Managers/CursorManager.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 20ba892711ef4d7ca7f9032b5d742d60
+timeCreated: 1580412504
