+++ Assets/Scripts/UI/InteractTooltipOld.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: cbfe16ddeb7f42d2acf1b6810ec8df4a
+timeCreated: 1580413344
